Тесты и отчет Allure
